Add automatic refresh and change expiry to 100 articles.
It turns out that I already called the FeedGoal function cleanDbItems
which expires articles.  I have changed the preferences to delete
articles when there are more than 100 articles in a feed.

Refresh is hardcoded to update feeds that haven't been updated for
an hour.  This check is done whenever the main screen is shown.