Pull in FeedParser project in Weader
Ugly hack neccesary to make Android import the classes from the
jar file.  It ought to be enough to go into Java Build Path and
add FeedParser the Projects tab and then check FeedParser in the
Order and Exports tab.  The Android Dev Team says that it should
work that way starting with ADT 17:

http://stackoverflow.com/questions/9649156/how-to-include-a-java-project-in-an-android-project

But no, it still doesn't work in ADT 22.  So what we have to do
is to generate FeedParser.jar (something Eclipse does not have
direct support for so we have to use an external ant.xml to do
that and that doesn't work very well) which is imported on the
Libraries tab and then exported. That works somewhat but is ugly.