Replace FeedGoal with FeedParser with a new database model
This replaces FeedGoal with FeedParser and introduces a new data model.

The big change in the databas is that the Article table (similar to
the Item table in FeedGoal) no longer holds the content for each
article.  By moving the rather big content column to a separate table
this speeds up other database operations that don't need the content,
such as showing the feed list with the number of unread articles
in each feed.

Also, the content is no longer preparsed, the raw XML for the entry/item
is stored in the database, so if I decide that I want to interpret
data from an article in a different way in the future, I can do that
instead of being stuck with what I stored in the database when the
article was first fetched.